Move the code for collecting supported RAID levels
Move the code for collecting supported RAID levels to interactive utils.
