+++ pyanaconda/modules/storage/partitioning/interactive_utils.py
@@ -20,7 +20,7 @@
 import re
 
 from blivet import devicefactory
-from blivet.devicelibs import crypto
+from blivet.devicelibs import crypto, raid
 from blivet.devices import LUKSDevice
 from blivet.errors import StorageError
 from blivet.formats import get_format
@@ -37,7 +37,6 @@ from pyanaconda.storage.root import Root
 from pyanaconda.storage.utils import filter_unsupported_disklabel_devices, bound_size, \
     get_supported_filesystems, PARTITION_ONLY_FORMAT_TYPES, SUPPORTED_DEVICE_TYPES, \
     CONTAINER_DEVICE_TYPES, DEVICE_TEXT_MAP
-from pyanaconda.ui.gui.spokes.lib.custom_storage_helpers import get_supported_raid_levels
 
 log = get_module_logger(__name__)
 
@@ -975,3 +974,27 @@ def collect_containers(storage, device_type):
         return storage.btrfs_volumes
     else:
         return storage.vgs
+
+
+def get_supported_raid_levels(device_type):
+    """Get RAID levels supported for the given device type.
+
+    It supports any RAID levels that it expects to support and that blivet
+    supports for the given device type.
+
+    Since anaconda only ever allows the user to choose RAID levels for
+    device type DEVICE_TYPE_MD, hiding the RAID menu for all other device
+    types, the function only returns a non-empty set for this device type.
+    If this changes, then so should this function, but at this time it
+    is not clear what RAID levels should be offered for other device types.
+
+    :param int device_type: one of an enumeration of device types
+    :return: a set of supported raid levels
+    :rtype: a set of instances of blivet.devicelibs.raid.RAIDLevel
+    """
+    if device_type == devicefactory.DEVICE_TYPE_MD:
+        supported = set(raid.RAIDLevels(["raid0", "raid1", "raid4", "raid5", "raid6", "raid10"]))
+    else:
+        supported = set()
+
+    return devicefactory.get_supported_raid_levels(device_type).intersection(supported)
